moved files to example folder
