+++ examples/levelset/levelset_old.py
@@ -1,22 +1,33 @@
 import firedrake as fd
 import fireshape as fs
+import fireshape.zoo as fsz
 import ROL
-from levelsetfunctional import LevelsetFunctional
 
-#setup problem
-mesh = fd.UnitSquareMesh(30, 30)
+dim = 2
+if dim == 2:
+    n = 100
+    mesh = fd.UnitSquareMesh(n, n)
+else:
+    n = 30
+    mesh = fd.UnitCubeMesh(n, n, n)
+
 Q = fs.FeControlSpace(mesh)
 inner = fs.LaplaceInnerProduct(Q)
+
+mesh_m = Q.mesh_m
+
+if dim == 2:
+    (x, y) = fd.SpatialCoordinate(mesh_m)
+    f = (pow(x-0.5, 2))+pow(y-0.5, 2) - 2.
+else:
+    (x, y, z) = fd.SpatialCoordinate(mesh_m)
+    f = (pow(x-0.5, 2))+pow(y-0.5, 2)+pow(z-0.5, 2) - 2.
+
 q = fs.ControlVector(Q, inner)
-
-#save shape evolution in file domain.pvd
 out = fd.File("domain.pvd")
-cb = lambda: out.write(Q.mesh_m.coordinates)
+J = fsz.LevelsetFunctional(f, Q, cb=lambda: out.write(mesh_m.coordinates))
 
-#create objective functional
-J = LevelsetFunctional(Q, cb=cb)
 
-#ROL parameters
 params_dict = {
     'General': {
         'Secant': {'Type': 'Limited-Memory BFGS',
@@ -29,7 +40,7 @@ params_dict = {
         'Gradient Tolerance': 1e-4,
         'Relative Gradient Tolerance': 1e-3,
         'Step Tolerance': 1e-10, 'Relative Step Tolerance': 1e-10,
-        'Iteration Limit': 30}}
+        'Iteration Limit': 150}}
 
 params = ROL.ParameterList(params_dict, "Parameters")
 problem = ROL.OptimizationProblem(J, q)
